webapp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
markdown==3.7
pygments==2.19.0
pyyaml==6.0.2
python-frontmatter==1.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.18
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
